# Searcher: return detected language