--- a/cmd/searcher/internal/search/search_regex.go
+++ b/cmd/searcher/internal/search/search_regex.go
@@ -135,7 +135,8 @@ func regexSearch(
 				}
 
 				// Apply language filters
-				if !lm.Matches(f.Name, getContent) {
+				langMatch, lang := lm.Matches(f.Name, getContent)
+				if !langMatch {
 					filesSkipped.Inc()
 					continue
 				}
@@ -145,6 +146,7 @@ func regexSearch(
 				match := false
 				fm := protocol.FileMatch{
 					Path:     f.Name,
+					Language: lang,
 					LimitHit: false,
 				}
 
@@ -163,7 +165,7 @@ func regexSearch(
 						// find limit+1 matches so we know whether we hit the limit
 						var locs [][]int
 						match, locs = m.MatchesFile(l.fileMatchBuf, sender.Remaining()+1)
-						fm = locsToFileMatch(l.fileBuf, f.Name, locs, contextLines)
+						fm = locsToFileMatch(l.fileBuf, f.Name, lang, locs, contextLines)
 					}
 				}
 
@@ -263,10 +265,11 @@ func readAll(r io.Reader, b []byte) (int, error) {
 	}
 }
 
-func locsToFileMatch(fileBuf []byte, name string, locs [][]int, contextLines int32) protocol.FileMatch {
+func locsToFileMatch(fileBuf []byte, name string, lang string, locs [][]int, contextLines int32) protocol.FileMatch {
 	if len(locs) == 0 {
 		return protocol.FileMatch{
 			Path:     name,
+			Language: lang,
 			LimitHit: false,
 		}
 	}
@@ -275,6 +278,7 @@ func locsToFileMatch(fileBuf []byte, name string, locs [][]int, contextLines int
 	cms := chunksToMatches(fileBuf, chunks, contextLines)
 	return protocol.FileMatch{
 		Path:         name,
+		Language:     lang,
 		ChunkMatches: cms,
 		LimitHit:     false,
 	}