--- a/cmd/searcher/internal/search/search_regex.go
+++ b/cmd/searcher/internal/search/search_regex.go
@@ -166,7 +166,6 @@ func regexSearch(
 					if langMatch {
 						fm.Language = lang
 						sender.Send(fm)
-						continue
 					}
 				}
 			}