… internal/searcher/v1/searcher.pb.go