For searches that hit Zoekt or the symbols backend, we return the file's
detected language through `result.File.PreciseLanguage`. Now, searcher will
also return the detected language whenever it's available. This value is used
in the search results filters panel, so it's important to be as accurate as
possible.

The precise language is only available when these criteria are met:
* The `search-content-based-lang-detection` feature is enabled -- we hope to
enable this by default soon
* The search contains a language filter, so we've already run language
detection

The PR also adds an optimization to run the expensive language detection
only after checking the search pattern.

Relates to https://github.com/sourcegraph/sourcegraph/issues/60341

## Test plan

Updated unit tests, manual testing