@jtibshirani 2024-02-19
This is a big improvement, but still leaves some cases where we don't have the precise language. Here's an example where the filters panel can still be inaccurate (this is a C++ only repo):
![Screenshot 2024-02-19 at 1 58 15 PM](https://github.com/sourcegraph/sourcegraph/assets/7461306/bb4580ba-b212-49d3-a646-a00e87bda134)
We have a few options:
1. Always run language detection on every match, even if there are no language filters. 
2. Accept the inconsistency.
3. Don't display language filters in the filters panel unless we have `PreciseLanguage` results. In practice, this means indexed and symbol searches are fine, but some unindexed searches are missing the lang filters in the panel.

I don't like option (1) because it would significantly slow down unindexed search in some common cases. Even for a `type:path` search, we would always have to load file contents, and always run go-enry. My preferred approach is (3). What do you think?