@camdencheek 2024-02-20
> it would significantly slow down unindexed search

Have you measured this? I was also corrected by @varungandhi-src on this, but the language detection only actually calls the "get the file contents" function if there is ambiguity, which there isn't for most file extensions. So I imagine this would impact the speed of C/C++ codebases, but not, for example, Go.